# backend/switchManagerAPI/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/v1/connections/":{"get":{"tags":["v1","connections"],"summary":"Listconnections","description":"return a paginated list of connections","operationId":"listConnections_api_v1_connections__get","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Page"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}},{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Search"}},{"name":"sort","in":"query","required":false,"schema":{"allOf":[{"$ref":"#/components/schemas/ListSortEnum"}],"default":"con","title":"Sort"}},{"name":"order","in":"query","required":false,"schema":{"allOf":[{"$ref":"#/components/schemas/OrderBy"}],"default":"asc","title":"Order"}},{"name":"filter","in":"query","required":false,"schema":{"allOf":[{"$ref":"#/components/schemas/ListFilterEnum"}],"default":"all","title":"Filter"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConnectionsOutput"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/connections/{id}":{"get":{"tags":["v1","connections"],"summary":"Getconnection","description":"return a connection","operationId":"getConnection_api_v1_connections__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConnectionOutput"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/connections/upsert":{"post":{"tags":["v1","connections"],"summary":"Upsertconnection","description":"upsert or udpate one || multiple connections","operationId":"upsertConnection_api_v1_connections_upsert_post","requestBody":{"content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/ConnectionUpsertInput"},{"items":{"$ref":"#/components/schemas/ConnectionUpsertInput"},"type":"array"}],"title":"Input"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/switchmanagerapi__models__factories__batcheableOutputFactory___locals___BatcheableOutputModel__1"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/connections/delete/{id}":{"delete":{"tags":["v1","connections"],"summary":"Deleteconnection","description":"delete a connection","operationId":"deleteConnection_api_v1_connections_delete__id__delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string","title":"Response Deleteconnection Api V1 Connections Delete  Id  Delete"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/customers/":{"get":{"tags":["v1","customers"],"summary":"Listcustomers","description":"return a list of customers","operationId":"listCustomers_api_v1_customers__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Customer"},"type":"array","title":"Response Listcustomers Api V1 Customers  Get"}}}},"404":{"description":"Not found"}}}},"/api/v1/customers/{id}":{"get":{"tags":["v1","customers"],"summary":"Getcustomer","description":"return a customer","operationId":"getCustomer_api_v1_customers__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Customer"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/customers/upsert":{"post":{"tags":["v1","customers"],"summary":"Upsertcustomer","description":"upsert or udpate one || multiple customer(s)","operationId":"upsertCustomer_api_v1_customers_upsert_post","requestBody":{"content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/UpsertCustomerInput"},{"items":{"$ref":"#/components/schemas/UpsertCustomerInput"},"type":"array"}],"title":"Input"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/switchmanagerapi__models__factories__batcheableOutputFactory___locals___BatcheableOutputModel__2"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/customers/delete/{id}":{"delete":{"tags":["v1","customers"],"summary":"Deletecustomer","description":"delete a customer","operationId":"deleteCustomer_api_v1_customers_delete__id__delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string","title":"Response Deletecustomer Api V1 Customers Delete  Id  Delete"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/switches/":{"get":{"tags":["v1","switches"],"summary":"Listswitches","description":"return a list of switches","operationId":"listSwitches_api_v1_switches__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Switch"},"type":"array","title":"Response Listswitches Api V1 Switches  Get"}}}},"404":{"description":"Not found"}}}},"/api/v1/switches/{id}":{"get":{"tags":["v1","switches"],"summary":"Getswitch","description":"return a switch","operationId":"getSwitch_api_v1_switches__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Switch"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/switches/upsert":{"post":{"tags":["v1","switches"],"summary":"Upsertswitch","description":"upsert or udpate one || multiple switch(s)","operationId":"upsertSwitch_api_v1_switches_upsert_post","requestBody":{"content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/UpsertSwitchInput"},{"items":{"$ref":"#/components/schemas/UpsertSwitchInput"},"type":"array"}],"title":"Input"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/switchmanagerapi__models__factories__batcheableOutputFactory___locals___BatcheableOutputModel__3"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/switches/delete/{id}":{"delete":{"tags":["v1","switches"],"summary":"Deleteswitch","description":"delete a switch","operationId":"deleteSwitch_api_v1_switches_delete__id__delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string","title":"Response Deleteswitch Api V1 Switches Delete  Id  Delete"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BatchError":{"properties":{"id":{"type":"string","title":"Id","default":""},"error":{"type":"string","title":"Error","default":""}},"type":"object","title":"BatchError","description":"Batch error model"},"ConnectionOutput":{"properties":{"id":{"type":"string","maxLength":255,"minLength":3,"title":"Id","description":"connection id"},"ppp":{"type":"string","maxLength":255,"minLength":3,"title":"Ppp","description":"connection name"},"port":{"type":"integer","title":"Port","description":"port number on the switch","default":0,"max":65535,"min":0},"toggled":{"type":"boolean","title":"Toggled","description":"define if the port is opened","default":false},"toggleDate":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Toggledate","description":"date at which the port should open / close based on currrent port status"},"type":{"type":"string","title":"Type","description":"physical connection type","default":"copper|fiber"},"isUp":{"type":"boolean","title":"Isup","default":false},"proto":{"type":"string","title":"Proto","default":"snmp"},"speed":{"type":"integer","title":"Speed","default":0},"switch":{"allOf":[{"$ref":"#/components/schemas/Switch"}]},"customer":{"allOf":[{"$ref":"#/components/schemas/Customer"}]}},"type":"object","required":["id","ppp"],"title":"ConnectionOutput","description":"Connection model API output"},"ConnectionUpsertInput":{"properties":{"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"port":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Port"},"toggled":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Toggled"},"toggleDate":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Toggledate"},"type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Type"},"isUp":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Isup"},"proto":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Proto"},"speed":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Speed"},"switchId":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Switchid"},"customerId":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Customerid"}},"type":"object","title":"ConnectionUpsertInput","description":"Connection model upsert API input"},"ConnectionsOutput":{"properties":{"connections":{"items":{"$ref":"#/components/schemas/ConnectionOutput"},"type":"array","title":"Connections","default":[]},"hasPrevious":{"type":"boolean","title":"Hasprevious","default":false},"hasNext":{"type":"boolean","title":"Hasnext","default":false}},"type":"object","title":"ConnectionsOutput","description":"connections paginated output"},"Customer":{"properties":{"id":{"type":"string","maxLength":255,"minLength":3,"title":"Id","description":"customer id / worker id"},"firstname":{"type":"string","maxLength":255,"minLength":1,"title":"Firstname"},"lastname":{"type":"string","maxLength":255,"minLength":1,"title":"Lastname"},"type":{"type":"string","maxLength":255,"minLength":1,"title":"Type","description":"customer type (company name || person status)"},"address":{"type":"string","maxLength":255,"minLength":1,"title":"Address"}},"type":"object","required":["id","firstname","lastname","type","address"],"title":"Customer","description":"Customer Database model"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ListFilterEnum":{"type":"string","enum":["all","customer","address","enabled","disabled","up","down","port","switch"],"title":"ListFilterEnum","description":"to filter down connections search results"},"ListSortEnum":{"type":"string","enum":["con","name","cid","address","switch"],"title":"ListSortEnum"},"OrderBy":{"type":"string","enum":["asc","desc"],"title":"OrderBy","description":"to order results in a paginated request"},"Switch":{"properties":{"id":{"type":"string","maxLength":255,"minLength":3,"title":"Id","description":"switch unique name / id"},"ip":{"type":"string","maxLength":15,"minLength":7,"title":"Ip","description":"switch ip address","default":"0.0.0.0"},"gpsLat":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Gpslat","max":90,"min":-90},"gpsLong":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Gpslong","max":180,"min":-180}},"type":"object","required":["id"],"title":"Switch","description":"Switch Database model"},"UpsertCustomerInput":{"properties":{"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"firstname":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Firstname"},"lastname":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Lastname"},"type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Type"},"address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address"}},"type":"object","title":"UpsertCustomerInput"},"UpsertSwitchInput":{"properties":{"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"ip":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ip"},"gpsLat":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Gpslat"},"gpsLong":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Gpslong"}},"type":"object","title":"UpsertSwitchInput"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"switchmanagerapi__models__factories__batcheableOutputFactory___locals___BatcheableOutputModel__1":{"properties":{"items":{"items":{"$ref":"#/components/schemas/ConnectionOutput"},"type":"array","title":"Items","default":[]},"errors":{"items":{"$ref":"#/components/schemas/BatchError"},"type":"array","title":"Errors","default":[]}},"type":"object","title":"BatcheableOutputModel","description":"batcheable output model"},"switchmanagerapi__models__factories__batcheableOutputFactory___locals___BatcheableOutputModel__2":{"properties":{"items":{"items":{"$ref":"#/components/schemas/Customer"},"type":"array","title":"Items","default":[]},"errors":{"items":{"$ref":"#/components/schemas/BatchError"},"type":"array","title":"Errors","default":[]}},"type":"object","title":"BatcheableOutputModel","description":"batcheable output model"},"switchmanagerapi__models__factories__batcheableOutputFactory___locals___BatcheableOutputModel__3":{"properties":{"items":{"items":{"$ref":"#/components/schemas/Switch"},"type":"array","title":"Items","default":[]},"errors":{"items":{"$ref":"#/components/schemas/BatchError"},"type":"array","title":"Errors","default":[]}},"type":"object","title":"BatcheableOutputModel","description":"batcheable output model"}}}}